Add chain support
