+++ examples/ethereum/types.go
@@ -1,68 +1,11 @@
 package ethereum
 
 import (
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/golang/protobuf/ptypes/timestamp"
-	"github.com/rs/zerolog/log"
+	"google.golang.org/protobuf/proto"
 )
 
-func UnmarshalEthHeader(in *types.Header) *Block0 {
-	return &Block0{
-		Ts:         &timestamp.Timestamp{Seconds: int64(in.Time)},
-		Hash:       in.Hash().Bytes(),
-		Num:        in.Number.Uint64(),
-		Difficulty: in.Difficulty.Uint64(),
-		GasLimit:   in.GasLimit,
-		GasUsed:    in.GasUsed,
-		Nonce:      in.Nonce.Uint64(),
-	}
-}
-
-// go-ethereum Block to nakji struct
-func (x *Block) UnmarshalEthBlock(in *types.Block) {
-	*x = Block{
-		Ts:         &timestamp.Timestamp{Seconds: int64(in.Time())},
-		Hash:       in.Hash().Hex(),
-		Difficulty: in.Difficulty().Uint64(),
-		Number:     in.Number().Uint64(),
-		GasLimit:   in.GasLimit(),
-		GasUsed:    in.GasUsed(),
-		Nonce:      in.Nonce(),
-	}
-}
-
-// go-ethereum Transaction to nakji struct
-func (tx *Transaction) UnmarshalEthTransaction(in *types.Transaction) {
-	V, R, S := in.RawSignatureValues()
-
-	// handle nil recipients for contract creations
-	Recipient := []byte{}
-	if in.To() != nil {
-		Recipient = in.To().Bytes()
-	}
-
-	// Get Sender (.From()) address
-	from, err := types.Sender(types.LatestSignerForChainID(in.ChainId()), in)
-	if err != nil {
-		log.Fatal().Err(err).
-			Interface("tx", in).
-			Msg("UnmarshallEthTransaction .AsMessage error")
-	}
-
-	*tx = Transaction{
-		AccountNonce: in.Nonce(),
-		Price:        in.GasPrice().Uint64(),
-		GasLimit:     in.Gas(),
-		Recipient:    Recipient, // nil means contract creation
-		Amount:       in.Value().Uint64(),
-		Payload:      in.Data(),
-
-		From: from.Bytes(),
-		Hash: in.Hash().Hex(),
-		Size: float64(in.Size()),
-
-		V: V.Uint64(),
-		R: R.Uint64(),
-		S: S.Uint64(),
-	}
+var protos = []proto.Message{
+	&Block{},
+	&Transaction{},
+	&Block0{},
 }
